IO Forwarding for Tools and Debuggers
This RFC extends support for tools and debuggers by adding the ability for the RM to forward output from applications to a tool, and to forward stdin from the tool to one or more application processes. Enabling this requires the addition of several new attributes and APIs covering the cases where an application is being spawned using the PMIx_Spawn API versus where the tool is connecting to an already-running application.

Signed-off-by: Ralph Castain <rhc@open-mpi.org>