Acquisition of Subsystem Launch Information
The PMIx launch procedure includes a stage whereby various subsystems (e.g., fabric) can provide a "blob" to the resource manager that is to be delivered to the PMIx server on each compute node. The blob can comprise a wide range of information, including propagation of environmental variables and fabric endpoint assignments.

Signed-off-by: Ralph Castain <rhc@open-mpi.org>