Extension of PMIx Logging Support
RFC0010 added a new PMIx_Log_nb API for requesting logging of provided data in some global data store or to standard output locations (e.g., stderr or syslog), subject to available services from the host environment. A limited number of initial keys were provided by which the logging request could direct the data to specific channels on an as-available basis, and the host RM interface was extended through the addition of a "log" interface by which client requests that are not directly supported by the PMIx library can be passed to the host RM for handling - e.g., when the logging data is directed to the stderr or stdout channels.

This RFC extends that capability by adding new logging channels along
with a blocking version of the PMIx_Log call

Signed-off-by: Ralph Castain <rhc@open-mpi.org>